Fixed verification flow and now use modals for success/failure
Added new images failure.png and success.png for use in the verify page
Fixed the checkin attestor to use the ES256K-R algorithm instead of ES256K to get rid of the '-1 is not property of null' bug.  Also added issue uport-project/did-jwt#14 documenting the wackness.
Made sure that the modal doesn't appear until after all verifications are tested.